cookbook/test_multiuser/requirements.txt
pocketflow
google-genai
sentence-transformers
faiss-cpu
numpy
pyyaml

